requirements.txt
absl-py==2.3.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.13.1
fonttools==4.58.0
fsspec==2024.6.1
grpcio==1.71.0
Jinja2==3.1.4
joblib==1.5.1
kiwisolver==1.4.8
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
numpy==2.1.2
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
protobuf==6.31.0
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.3
setuptools==70.2.0
six==1.17.0
standard-imghdr==3.13.0
sympy==1.13.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
torch==2.7.0+cu126
torchaudio==2.7.0+cu126
torchvision==0.22.0+cu126
typing_extensions==4.12.2
Werkzeug==3.1.3
